requirements.txt
torch==1.10.0
torchvision==0.11.1
pytorch-lightning==1.5.3
torchmetrics>=0.4.0
segmentation-models-pytorch>=0.2.0
hydra-core==1.1.1
kornia
albumentations
pandas
tensorboardX
pillow
matplotlib
scipy
numpy
pytorch_toolbelt
descartes
fiona
psycopg2
shapely==1.8.0
geopandas==0.10.2
geoalchemy2
rasterio
numba
sahi
skan==0.9.0
torch-scatter
tqdm
pygeos
rtree
bidict
Cython
ninja
pyyaml
pycocotools
multiprocess
wget
fastapi==0.70.0
uvicorn==0.15.0
similaritymeasures
